bitrix_helper/senderMessage/pyproject.toml
[tool.poetry]
name = "SIGNAL-bot"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
aiogram = "^3.8.0"
aiohttp = "^3.9.5"
requests = "^2.32.3"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
